refacotored version. fix installation bug between numpy and scipy. Adding of docker initial version. rename gnix app to ginxcli because of name conflict. we still use ginx as our cli.
